Revert use case 3
Although possible, it could be implemented in cleaner ways.
