+++ src/Common/EventStore/Integration/Predis/PredisEventStorePointerFactory.php
@@ -6,7 +6,6 @@ namespace Gaming\Common\EventStore\Integration\Predis;
 
 use Gaming\Common\EventStore\EventStorePointer;
 use Gaming\Common\EventStore\EventStorePointerFactory;
-use Gaming\Common\EventStore\InMemoryCacheEventStorePointer;
 use Predis\ClientInterface;
 
 final class PredisEventStorePointerFactory implements EventStorePointerFactory
@@ -18,11 +17,9 @@ final class PredisEventStorePointerFactory implements EventStorePointerFactory
 
     public function withName(string $name): EventStorePointer
     {
-        return new InMemoryCacheEventStorePointer(
-            new PredisEventStorePointer(
-                $this->predis,
-                $name
-            )
+        return new PredisEventStorePointer(
+            $this->predis,
+            $name
         );
     }
 }
